# [FIX] xlsx: accelerate xlsx export