--- a/src/xlsx/helpers/content_helpers.ts
+++ b/src/xlsx/helpers/content_helpers.ts
@@ -140,8 +140,9 @@ export function addRelsToFile(
 
 export function pushElement<T>(property: T, propertyList: T[]): number {
   let len = propertyList.length;
+  const operator = typeof property === "object" ? deepEquals : (a: T, b: T) => a === b;
   for (let i = 0; i < len; i++) {
-    if (deepEquals(property, propertyList[i])) {
+    if (operator(property, propertyList[i])) {
       return i;
     }
   }